Validate template and network policy contracts
Covers runtime network policy validation, shared template network validation, and removal of unsupported template spec fields from the public contract.
